--- src/components/Navbar/Navbar.module.css
@@ -1,3 +1,0 @@
-.test {
-  font-size: 50px;
-}
