+++ src/pages/DashboardPage/DashBoardPage.module.css
@@ -1,25 +1,3 @@
 /* .body {
   width: 1022px;
 } */
-
-.dashboardList {
-  width: 100%;
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr; /* 각 행의 높이 */
-  grid-template-rows: 1fr 1fr; /* 각 열의 너비 */
-  gap: 15px;
-}
-
-@media screen and (max-width: 744px) {
-  .dashboardList {
-    grid-template-columns: 1fr 1fr; /* 각 행의 높이 */
-    grid-template-rows: 1fr 1fr 1fr; /* 각 열의 너비 */
-  }
-}
-
-@media screen and (max-width: 375px) {
-  .dashboardList {
-    grid-template-columns: 1fr; /* 각 행의 높이 */
-    grid-template-rows: 1fr 1fr 1fr 1fr 1fr 1fr; /* 각 열의 너비 */
-  }
-}
